PYTHONCODES/32/motivation/pyproject.toml
[project]
name = "motivation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-decouple>=3.8",
]
